+++ src/plugins/kibana/public/settings/sections/kibi_datasources/services/_saved_datasource.js
@@ -24,7 +24,7 @@ define(function (require) {
           description: 'string',
           datasourceType: 'string',
           datasourceParams: 'json',
-          version: 'long'
+          version: 'integer'
         },
 
         defaults: {
